move test_partial_message_read
